frontend/worker/requirements.txt
rq==1.15.1
redis==5.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
